requirements.txt
flask
numpy
pillow
torch
transformers
faiss-cpu
